moved backup methods to BackupManager
