+++ Mapping Tools/views/SliderCompletionator/SliderCompletionatorView.xaml.cs
@@ -59,7 +59,7 @@ namespace Mapping_Tools.Views.SliderCompletionator {
             // Remove logical focus to trigger LostFocus on any fields that didn't yet update the ViewModel
             FocusManager.SetFocusedElement(FocusManager.GetFocusScope(this), null);
 
-            IOHelper.SaveMapBackup(paths);
+            BackupManager.SaveMapBackup(paths);
 
             ViewModel.Paths = paths;
             ViewModel.Quick = quick;
